+++ BonaLiz.Negocio/Utils/Formater.cs
@@ -4,7 +4,7 @@ namespace BonaLiz.Negocio.Helpers
 {
 	public static class Formater
 	{
-		public static string FormatarMoeda(decimal valor) {
+		public static string FormatarMoeda(double valor) {
 			return valor.ToString("C", CultureInfo.GetCultureInfo("pt-BR"));
 		}
 	}
